backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
deepface==0.0.79
google-api-python-client==2.108.0
pafy==0.5.5
yt-dlp==2023.10.7
starlette==0.27.0

